requirements.txt
requests
pandas
python-dotenv
openai
rapidfuzz
feedparser
gspread
oauth2client
pytest
tqdm